test file
This was an original draft for a project, me and my friend worked on it. The project was originally to design a menu for a Club to review events and select the ticket price for them. We included functions to calculate and keep track of the total price. So far it works well, with a minor register clearing error for BX in the Comp function, it was fixed with a simple command tho. Any ideas are welcome